Added Strategies. Added Parking lot service

+++ Parking Lot System/src/model/ticket/Receipt.java
@@ -5,10 +5,10 @@ import java.time.Instant;
 public class Receipt {
 
     private final Ticket ticket;
-    private final int amount;
+    private final long amount;
     private final Instant exitInstant;
 
-    public Receipt(Ticket ticket, int amount, Instant exitInstant) {
+    public Receipt(Ticket ticket, long amount, Instant exitInstant) {
         this.ticket = ticket;
         this.amount = amount;
         this.exitInstant = exitInstant;
@@ -18,7 +18,7 @@ public class Receipt {
         return ticket;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
